Adds possibility to search events by tags
This commit introduces the possibility to search for events based on
tags. The URI would then be /v2.1/events/?tags=<tag1>,<tag2>,<tag3>

The value given to the tags-keyword is a comma-separated list of tags.
The selected events will have at least one of those keywords.

Implemented is an OR-style tags-selection.