Adds searching events by date
This commit addresses JOINDIN-448 by adding a possibility to search
events by date.

Start- ans Enddate can be given as UTC-Timestamps or as any timestring
recognized by DateTime that will then be converted into an
UTC-Timestamp.

This search will select events whose begin is within the bounds of
start- and endtime. The end of the event is not checked currently.