Adds filter for tags and title
This commit adds a filter-mechanism that allows additional filtering
based on the existing filters to further reduce the number of searched
events.

Filtering can be done using additional parameters either in single or in
array notation like follows:

    events/?title[]=php&title[]=foo

This filters all events whos title contains either 'php' or 'foo'. For
an 'AND' selection you can add ```type=AND``` to the query.

Alternatively one can also filter by tags using the same notation:

    events/?tag[]=php&tag[]=usergroup

THis will select all events that are tagged with either 'php' or
'usergroup'. To select events tagged with 'php' AND 'usergroup' one will
have to add ```type=AND``` to the request.

To get ```AND``` connection of tags and ```OR```-connection for title
one can use ```&type[title]=OR&type[tag]=AND```

Default connection is ```OR```