Adds filtering for dates
This Commit Addresses #JOINDIN-448 and #JOINDIN-449

This commit adds the parameters ```startdate``` and ```enddate``` to the
filtering-options.

Date-filtering allows any string parseable by DateTime to be used as
input string. If no valid input string is given the current timerstamp will be
used.

It is possible to use multiple ```startdate``` and ```enddate```-
parameters even though it's use is debateable