Adds location-based filtering
For location based filtering three parameters are necessary:

* lat - the latitude of the center
* lng - The longitude of the center
* distance - The radius of the circle around the center.

All distances are measured in kilometers.

So to filter events in a diameter of 20 kilometers around 50° north/ 8°
west you would use ```events/?lat=50&lng=8&distance=20```

This of coures only works out for events that have a location set. All
events without a location will not be available when applying this
filter to any events-search.

This addresses #JOINDIN-449