Adds total count of events
Addresses #JOINDIN-397.

This commit introduces a second SQL-Query that counts all matching
events. That total count is then added to the returned value and also
used to check whether there is a next page to be displayed or not
based on the requested LIMIT-parameters.

This changes the returned value in two ways:

* A new parameter ```meta.total``` is added to the returned values
* On certain occasions the next-link is not present where it was present
  before.