#457: Event submit crashes when CFP Start/End is given
In this commit we fix a fatal error that occurs if a CFP start or end
date is given. The start and end date are converted to DateTime objects
but never back into strings before it is inserted in the database.

In addition the the CFP start and end date were not added onto the INSERT
query as prepared statements; causing the actual insert to fail as well.

Both issues are resolved with this commit.