Submitting an event should return 202 if pending
When an event is submitted by a user that does not make the event
automatically be approved then the returned status should be 202
and no location header should be returned.

Included in this commit is that the location header for an approved
event (hence a 201) is the complete URL and not just that of the event.