JOINDIN-475 Use Forwarded and X-Forwarded-For headers in Akismet check
Add new method to Request object to set the user's IP address based on whether or not there is a X-Forwarded-For is present.

Also, refactor the SpamCheckService to have the ip address and user agent passed in rather than pulling them directly from _SERVER. This included updating the two usages of SpamCheckService::isCommentAcceptable to pass in the ip address and user agent.