Jira 482 - Password Change API ==============================
Added Password controller that allows changing of password via the API.
The password will be updated by the supplied one only if the token is
correct and the client id and key indicate the client has grant privileges
for passwords. The API returns a new token as if the user had logged out
and logged back in again.