Removes Speaker-mamangement from the talks
THis commit removes management of speakers from the talks-controller to
the talks_speakercontroller.

To add a speaker to a talk you POST the speakers name to
"talks/[ID]/speakers" and you will get the URL of the speaker. To edit
or delete that speaker you PUT or DELETE to "speakers/[ID]"