Implement /facebook/token to log a user in
Take the verification code from the client, send to Facebook to get an access token.
With the access token, read the user's profile to get their email address. From that,
look up who they are, create them a new access token and return which is converted to
an access token from which we read the user's email address and then attempt to log
them in.