Adds tests and missing code for speaker-fetching
This commit adds frisby tests to check the speaker-listing for a talks
as well as the listing of a single speaker.

It also introduces methods needed to make the result of listing the
speakers for a talk reasonable. That means adding methods to show
informations for a single speaker by calling
http://api.joind.in/v2.1/speakers/<speaker-id>