JOINDIN-590 - Store inflected title in url_friendly_talk_title field
If there are no duplicated talk names in an event
then it will store the inflected but not decorated
version (no numeric suffix). If there are more
than one talk in an event with the same name, it
will try to store first without a suffix. If that
has already been taken then it will try with the
talk ID as a suffix. If that doesn't work either
then it is counted as a failure and will return
false. That should rarely, if ever, happen.