Update frisby write tests for talks
The frisby write tests need to use an access token for a user that we've
created during the run. To avoid creating another user, we use the
access token from the event tests and call runTalkTests() after we've
retrieved a newly created event.

This has the benefit that we can refactor the talk tests to use the
event's talks_uri property and so the talks tests do not need to
manually construct the URI to post to.

Finally, as the talks tests are now called within api_write_events.js,
we don't need to explicitly call them api_write_spec and can remove the
set up code from api_write_talks.js.