Extract generic features to AbstractModel
The constructor, __get() and the first part of getOuputView() are common
across all models, so place in a parent class. getOutputView() depends
on getDefaultFields() and getVerboseFields() which are declared abstract
as all children need to implement. However getSubResources() is declared
directly as some models do not have sub resources and so would not need
to implement this method.