Adds a way to retrieve speakers for a talk
This commit adds a way to get the speaker(s) for a talk via the API.
That way a list of speakers can be retreived for a single talk without
having to fetch the complete talk inforamtions.

It is also a prerequisite to add, edit or delete speakers for a talk
which in turn is a prerequisite to edit talks.

This commit introduces two new API-Endpoints:

* /talks(/[^/]+)*/speakers/?$ to get a list of all the speakers for a
  single talk and
* /speakers(/[^/]+)*/?$ to get information about a single speaker. This
  endpoint currently returns an array always containing a single entry.
  That's the same behaviour as the /talks(/[^/]+)*$-endpoint shows.

A speaker can be either a name if the speaker didn't claim the talk or a
user when the speaker did claim the talk. So there might be a user-URI
on a speaker-entry but that depends on whether the user claimed the talk
or not. And a speaker is unique to the talk. So one user doing two talks
can be found with two different speaker-entries.

There currently aren't frisby-tests. I'll add them in a follow-up PR
later.