Changes more 400 to 401
This commit changes existing 400-Exception codes to 401 as well as it
adds some 401-Exception codes where no code was set.

That way all exceptions that are thrown due to a user not being logged
in are now emitting an HTTP-Status code of 401.

There are als othree comments added where no Check for a logged in user
is in place. At those parts there is a later check for a user being able
to administer something and at least there the action is secured but it
might be easier to fail fast and check whether there'S a login and then
throw a 401