Adds prerequisits to ease unit-testing controllers
This change adds ways to omit the current calls to "header" and "exit"
in the controllers to send certain headers or to circumvent the output
of data to the client.

This change only renders output when content is provided, so when a
controller-action returns "null" no output is rendered.

Additionally headers can now be set using a "setHeader"-method on the
view, so direct calls to "header" aren't necessary any more