Fixes frisby tests
Due to database-differences between local and test-databases a user
isn't necessarily available in the database. So instead of assuming there
is a user this PR adds a database-retrieval of a user (with a random ID
though)

This issue popped up on merging the fix for the headers and should now
be resolved. There wasn't a ticket available