Shows only tokens that can be revoked in API
This commit causes that only tokens are retrieveable via the API that a
user should be able to remove. Tokens of some apps (like web2) should
not be revokable by the user. Such apps scan be marked in the
oauth_consumers-table by setting the value of the new field
can_be_revoked to 0.