JOINDIN-730 Clicking claim twice
This will be a breaking change if anywhere else approves claims apart
from web2.

The button on the talk page calls the same end point with the same data
as the admin one.

The count of the call is the differenciator.

This PR adds some data in the call to know which place you came from