Adds composer-execution to the launchscript
This commit adds the execution of composer to the script that launches a
new version of the joind.in-API.

Due to that it is possible to include external libraries only via adding
them to the composer.json (and therefore composer.lock)

As the script calls `composer install` only libraries and their versions
defined within the composer.lock will be installed!

It should be as safe as possible, as only libraries from the
composer.lock will be installed and those should have been tested
already. And should a library not be available that'd break the build
process so no build would go live under such circumstances.