Fixes changed users name on editing talks
As soon as a user changed the name and then edited a talk they where
already assigned to the new name would appear in the speaker-name field
but on submitting the already claimed talk was reassigned to the user
with the new name. That happened due to the API comparing the users name
to the value stored in the talk_claim table that might not be valid
anymore. Therefore that value no longer is checked on editing a talk.

This solves JOINDIN-743 and JOINDIN-745