Allow users to create a login via twitter
This change allows users to not only log in via twitter but also to
create a new account via login in from twitter.

This should make it much easier to get users to log into joind.in as
there's much less effort to be taken to get an account.

That should raise the adopition rate for conference-attendees.