Add keyword-based search to users endpoint
This will allow for improved search on the website. For instance an
user will be able to look up talks for a speaker they notice on twitter.
Additionally event organisers will be able to quickly navigate to the
user profile of a speaker in order to see previous talks or events.