Add a GET route to retrieve all images
- Created the GET route for /events/{id}/images
- Made the method `getImages` in the EventMapper model public so it can
be used for event data as well as images only
- Implemented an action method to return the list of images

Note that due to the implementation of `getImages` only one image of each
type is returned. We could change this to return a list of images per
type, but that will break API consumers. We also could make the images of
`/events/{id}` and `/events/{id}/images` have different data, but that
would not make sense.