Switch previously implicitly declared properties to public visibility
This resolves the bugged behavior from earlier today. Specifically:

1. As part of code cleanup to hit PHPStan Level 1, #735 explicitly
declared properties that had been implicitly set before. One catch:
implicit properties are public, but this declaration was private.
2. Talk link building relied on the talk ID property (one of the
implicitly set ones) to be public. When it wasn't, links that would
have included talk ID included '' instead. For example,
/talks//comments instead of /talks/1234/comments.
3. web2 used the talk comments link handed back from the API to grab
talk comments on talk page load. Since this was not, in fact, a valid
URL (even after some middlebox along the way dropped the extra slash),
the API threw back a 405.
4. web2, which assumes that the talk comments endpoint will always
return successfully, tried to grab the comments element from what was
actually a 405 error response, then iterate over that nonexistent value
to show comments, fatal'ing in the process.

So, while a bunch of pieces in this chain were brittle, the quick fix
for now is to switch properties back to being public, matching their
behavior from implicit declaration. Creating a follow-on issue to get
the properties back to private once we've made sure that every external
or descendant consumer of these properties does the consumption via a
getter (or updates via a setter).