Let model property accesses fall back to magic getter
So, I mis-diagnosed part of what was going on in #747.

While things were failing due to explicitly declared variables, they
were failing not because they had been implicitly declared before (for
models, anyway) but because models use __get to grab properties from an
internal array...even for accesses within the model (e.g. for building
links).

As a result, declaring properties, private or otherwise, "shadowed" the
magic getter that would've been called for those same properties,
creating the issue that we saw. So the solution here is to let the
magic getter do its job (though we may add explicit getters later) by
dropping the explicit properties on these models.