Evaluate more-specific talk comment report/moderation routes ahead of less-specific talk comment update routes
This fixes talk comment report moderation being shadowed by talk comment updates due to an overly-wide regex for the latter. One could argue that the proper way to fix this is by fixing the regex for the latter, but I'm not comfortable enough with regex-fu to attempt that, and since our router evaluates routes one by one, in order, this will do the trick.

Technically we don't need to move *both* report routes, as POST has no routes shadowing it, but having two talk comment routes separated in the code looks weird enough that someone might attempt to reorder the routes back, causing the bug to resurface.